requirements.txt
asgiref==3.5.2
attrs==23.1.0
bcrypt==4.0.1
black==23.3.0
blessed==1.20.0
boto3==1.26.121
botocore==1.29.121
cached-property==1.5.2
cachetools==5.3.0
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.3
cryptography==41.0.1
distlib==0.3.6
Django==4.1.3
django-crispy-forms==1.14.0
django-ses==3.4.1
django-tempus-dominus==5.1.2.17
Django-Verify-Email==2.0.3
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
filelock==3.12.0
flake8==6.0.0
flake8-black==0.3.6
idna==3.4
jmespath==1.0.1
jsonschema==3.2.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
paramiko==3.2.0
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
psycopg2==2.9.5
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyNaCl==1.5.0
pyproject_api==1.5.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.4.1
requests==2.28.2
s3transfer==0.6.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.3
termcolor==1.1.0
texttable==1.6.7
tomli==2.0.1
tox==3.28.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.23.0
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==6.4.0
